Improve performance test script
Includes adding the duration of the tests to the result, using ujson
to ensure we spend as little time as possible parsing data and so can
receive as quickly as possible, using uuid identifiers for subscriptions
and making the output look better.

The uuid identifiers remove potential issues in Strawberry of multiple
clients using the same ID for each subscription - previously it was
simply the PV's name. Real clients, e.g. Apollo, do something similar
to uuids.